backend/requirements.txt
django==5.0
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-crontab
mysqlclient
python-dotenv
pillow